refactor sample load cache start ownership
